# hotstock-be-v1/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{id}": {
      "delete": {
        "description": "Xóa vĩnh viễn một tài khoản người dùng.",
        "operationId": "UsersController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tài khoản đã bị xóa"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa tài khoản (Admin)",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "description": "Xem thông tin chi tiết một người dùng.",
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Chi tiết người dùng (Admin)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users": {
      "get": {
        "description": "Lấy danh sách người dùng, phân trang và lọc.",
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "user",
                "admin",
                "editor"
              ]
            }
          },
          {
            "name": "blocked",
            "required": false,
            "in": "query",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUsersDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Danh sách người dùng (Admin)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/me": {
      "get": {
        "description": "Lấy thông tin profile của user đang đăng nhập.",
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy thông tin cá nhân",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "description": "User tự cập nhật tên người dùng của mình.",
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật thông tin cá nhân",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/role": {
      "patch": {
        "description": "Thay đổi role của người dùng. Không thể tự đổi role chính mình.",
        "operationId": "UsersController_updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật phân quyền người dùng (Admin)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/plan": {
      "patch": {
        "description": "Chỉ định thủ công một gói trả phí cho người dùng.",
        "operationId": "UsersController_assignPlan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignPlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Gán gói cho người dùng (Admin)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/block": {
      "patch": {
        "description": "Khóa người dùng và buộc đăng xuất.",
        "operationId": "UsersController_blockUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tài khoản đã bị khóa"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Khóa tài khoản (Admin)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/unblock": {
      "patch": {
        "description": "Mở khóa cho người dùng đã bị khóa.",
        "operationId": "UsersController_unblockUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tài khoản đã được mở khóa"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mở khóa tài khoản (Admin)",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/tags": {
      "get": {
        "operationId": "TagsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ]
      },
      "post": {
        "operationId": "TagsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ]
      }
    },
    "/tags/{slug}": {
      "patch": {
        "operationId": "TagsController_update",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ]
      },
      "delete": {
        "operationId": "TagsController_remove",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ]
      }
    }
  },
  "info": {
    "title": "HotStock API Documentation",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserPlanDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "level": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "level"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "fullName": {
            "type": "object"
          },
          "phoneNumber": {
            "type": "object"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin",
              "editor"
            ]
          },
          "provider": {
            "type": "string"
          },
          "confirmed": {
            "type": "boolean"
          },
          "blocked": {
            "type": "boolean"
          },
          "passwordChangedAt": {
            "type": "object"
          },
          "plan": {
            "$ref": "#/components/schemas/UserPlanDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "role",
          "provider",
          "confirmed",
          "blocked",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedUsersDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "johndoe123",
            "description": "Tên người dùng mới"
          }
        }
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin",
              "editor"
            ],
            "description": "Role mới"
          }
        },
        "required": [
          "role"
        ]
      },
      "AssignPlanDto": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "number",
            "example": 1,
            "description": "ID của gói cần gán"
          }
        },
        "required": [
          "planId"
        ]
      },
...
      "CreateTagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Đầu tư giá trị",
            "description": "Tên thẻ"
          },
          "slug": {
            "type": "string",
            "example": "dau-tu-gia-tri"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateTagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Đầu tư giá trị",
            "description": "Tên thẻ"
          },
          "slug": {
            "type": "string",
            "example": "dau-tu-gia-tri"
          }
        }
      }
    }
  }
}